# 3 curs/1_half/server_app/openapi.yaml
openapi: 3.0.3
info:
  title: Аренда недвижимости
  description: Лабораторная работа №2, вариант 14
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/properties:
    get:
      summary: Получить все объекты недвижимости
      responses:
        '200':
          description: Список объектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
    post:
      summary: Добавить объект
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyRequest'
      responses:
        '201':
          description: Объект создан
          headers:
            Location:
              description: URI созданного объекта
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/properties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Получить объект по ID
      responses:
        '200':
          description: Объект найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Полностью обновить объект
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyRequest'
      responses:
        '200':
          description: Объект обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Удалить объект
      responses:
        '204':
          description: Объект удалён
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    PropertyRequest:
      type: object
      required: [address, type, area, rooms, monthlyRent, available]
      properties:
        address:
          type: string
          maxLength: 255
          example: Санкт-Петербург, Литейный проспект, 10
        type:
          type: string
          enum: [APARTMENT, HOUSE, ROOM, COMMERCIAL]
        area:
          type: number
          minimum: 0.1
          example: 42.5
        rooms:
          type: integer
          minimum: 1
          example: 2
        monthlyRent:
          type: number
          exclusiveMinimum: true
          minimum: 0
          example: 55000
        available:
          type: boolean
          example: true
    Property:
      allOf:
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              readOnly: true
        - $ref: '#/components/schemas/PropertyRequest'
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
  responses:
    BadRequest:
      description: Некорректные данные
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'